GUVNOR-2173,JBPM-5269 - Adding the jbpm-bpmn2-services module
