+++ jbpm-services/jbpm-bpmn2-services/.gitignore
@@ -1,0 +1,11 @@
+/target
+/local
+
+# Eclipse, Netbeans and IntelliJ files
+/.*
+!.gitignore
+/nbproject
+/*.ipr
+/*.iws
+/*.iml
+
