+++ jbpm-services/jbpm-bpmn2-services/src/main/resources/META-INF/services/org.jbpm.compiler.xml.ProcessDataEventListenerProvider
@@ -1,0 +1,1 @@
+org.jbpm.bpmn2.services.listener.ServicesProcessDataEventListenerProvider
