+++ jbpm-services/jbpm-kie-services/pom.xml
@@ -73,6 +73,10 @@
       <groupId>org.jbpm</groupId>
       <artifactId>jbpm-services-api</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.jbpm</groupId>
+      <artifactId>jbpm-bpmn2-services</artifactId>
+    </dependency>
 
     <!-- kie -->
     <dependency>
