--- jbpm-services/jbpm-kie-services/src/main/resources/META-INF/services/org.jbpm.compiler.xml.ProcessDataEventListenerProvider
@@ -1,1 +1,0 @@
-org.jbpm.kie.services.impl.xml.ServicesProcessDataEventListenerProvider
