+++ jbpm-services/jbpm-services-cdi/src/main/java/org/jbpm/services/cdi/impl/BPMN2DataServiceCDIImpl.java
@@ -19,7 +19,7 @@ package org.jbpm.services.cdi.impl;
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.event.Observes;
 
-import org.jbpm.kie.services.impl.bpmn2.BPMN2DataServiceImpl;
+import org.jbpm.kie.services.impl.bpmn2.BPMN2DeploymentDataServiceImpl;
 import org.jbpm.services.api.DeploymentEvent;
 import org.jbpm.services.cdi.Activate;
 import org.jbpm.services.cdi.Deactivate;
@@ -27,13 +27,13 @@ import org.jbpm.services.cdi.Deploy;
 import org.jbpm.services.cdi.Undeploy;
 
 @ApplicationScoped
-public class BPMN2DataServiceCDIImpl extends BPMN2DataServiceImpl {
+public class BPMN2DataServiceCDIImpl extends BPMN2DeploymentDataServiceImpl {
 
 	@Override
     public void onDeploy(@Observes@Deploy DeploymentEvent event) {
         super.onDeploy(event);
     }
-    
+
 	@Override
     public void onUnDeploy(@Observes@Undeploy DeploymentEvent event) {
         super.onUnDeploy(event);
